pyproject.toml
[tool.poetry]
name = "fetch-pubmed"
version = "0.1.0"
description = "Fetch papers from PubMed and filter by affiliation."
authors = ["md kaif"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.14" 
biopython = "^1.81"
pandas = "^2.2.3"


[tool.poetry.scripts]
get-papers-list = "fetch_pubmed:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
